Improve valves configuration readability

+++ include/control/valves/devices.hpp
@@ -4,10 +4,10 @@
     Base servo valves configuration
 */
 
-#define NUMBER_OF_VALVES 4
-#define SERVOS_OPEN_ANGLE 0
-#define SERVOS_CLOSE_ANGLE 180
-#define SERVOS_OPEN_CLOSE_TIMEOUT 3000
+#define NUMBER_OF_VALVES (unsigned short) 4
+#define SERVOS_OPEN_ANGLE (unsigned short) 0
+#define SERVOS_CLOSE_ANGLE (unsigned short) 180
+#define SERVOS_OPEN_CLOSE_TIMEOUT (unsigned short) 3000
 
 /*
     Servo valves name
@@ -16,27 +16,27 @@
     in Modules.hpp file
 */
 
-#define unknownValve -1
-#define aquariumWaterValve 0
-#define addionalWaterTankValve 1
-#define cleanWaterValve 2
-#define sewageWaterValve 3
+#define unknownValve (short) -1
+#define aquariumWaterValve (short) 0
+#define addionalWaterTankValve (short) 1
+#define cleanWaterValve (short) 2
+#define sewageWaterValve (short) 3
 
 /*
     Valves servos pwm pinout definition
 */
 
-#define AQUARIUM_WATER_VALVE_PIN 3
-#define ADDIONAL_WATER_TANK_VALVE_PIN 4
-#define CLEAN_WATER_VALVE_PIN 5
-#define SEWAGE_WATER_VALVE_PIN 6
-#define REMOTES_DETECTION_PIN 40
+#define AQUARIUM_WATER_VALVE_PIN (unsigned short) 3
+#define ADDIONAL_WATER_TANK_VALVE_PIN (unsigned short) 4
+#define CLEAN_WATER_VALVE_PIN (unsigned short) 5
+#define SEWAGE_WATER_VALVE_PIN (unsigned short) 6
+#define REMOTES_DETECTION_PIN (unsigned short) 40
 
 /*
     Valves close detection pinout definition
 */
 
-#define AQUARIUM_WATER_VALVE_CLOSE_DETECTION_PIN 33
-#define ADDIONAL_WATER_TANK_VALVE_CLOSE_DETECTION_PIN 35
-#define CLEAN_WATER_VALVE_CLOSE_DETECTION_PIN 37
-#define SEWAGE_WATER_VALVE_CLOSE_DETECTION_PIN 39
+#define AQUARIUM_WATER_VALVE_CLOSE_DETECTION_PIN (unsigned short) 33
+#define ADDIONAL_WATER_TANK_VALVE_CLOSE_DETECTION_PIN (unsigned short) 35
+#define CLEAN_WATER_VALVE_CLOSE_DETECTION_PIN (unsigned short) 37
+#define SEWAGE_WATER_VALVE_CLOSE_DETECTION_PIN (unsigned short) 39
